+++ public/index.js
@@ -22,13 +22,13 @@ exports.WaterfallRenderContext = require('./WaterfallRenderContext.js');
  * @param {...Promise<*>} promises Promises that resolve once loading data has been cached. The values resolved don’t matter. Multiple arguments can be used, similar to how [`Array.push`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/push) works.
  * @example <caption>Loading data in a [React](https://reactjs.org) component within a server and client side rendered app.</caption>
  * ```jsx
- * import { useContext } from 'react';
+ * import React from 'react';
  * import WaterfallRenderContext from 'react-waterfall-render/public/WaterfallRenderContext.js';
  * import useUserProfileData from '../hooks/useUserProfileData.mjs';
  * import UserProfile from './UserProfile.mjs';
  *
  * export default function UserPage({ userId }) {
- *   const declareLoading = useContext(WaterfallRenderContext);
+ *   const declareLoading = React.useContext(WaterfallRenderContext);
  *   const { load, loading, cache } = useUserProfileData(userId);
  *
  *   // For this example, assume loading errors are cached.