+++ changelog.md
@@ -6,10 +6,12 @@
 
 - Updated Node.js support to `^12.20.0 || ^14.13.1 || >= 16.0.0`.
 - Updated dev dependencies, some of which require newer Node.js versions than previously supported.
+- Now uses `React.createElement` instead of the [the new React JSX runtime](https://reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html).
 
 ### Patch
 
 - Also run GitHub Actions CI with Node.js v17.
+- Avoid named imports from [`react`](https://npm.im/react) and [`react-dom`](https://npm.im/react-dom) as they’re not proper Node.js ESM.
 - Renamed imports in the test index module.
 - Test the bundle sizes for public modules individually.
 - Amended the changelog entry for v2.0.0.